Moved ConnectableDevicePairingLevel static property to the DeviceService class https://github.com/ConnectSDK/Connect-SDK-Android/issues/121
